+++ src/resources/boards/board.model.ts
@@ -4,8 +4,8 @@ import {Column} from './column.model';
 export class Board {
   /**
    * Board class represents a custom structure of in-memory database instance
-   * The board represents an object that contains the columns and has a title
-   * Every board has its unique identificator created by v4 uuid
+   * The boards represents an object that contains the columns and has a title
+   * Every boards has its unique identificator created by v4 uuid
    */
   id: string;
