+++ src/resources/tasks/task.model.ts
@@ -3,9 +3,9 @@ import { v4 as uuid } from 'uuid';
 export class Task {
   /**
    * Task class represents a custom structure of in-memory database instance
-   * The task represents an object that contains a title, description and stores 
-   * a user, board and column links using their unique identificators
-   * Every task has its unique identificator created by v4 uuid
+   * The tasks represents an object that contains a title, description and stores
+   * a users, boards and column links using their unique identificators
+   * Every tasks has its unique identificator created by v4 uuid
    */
   id: string;
